agent/requirements.txt
requests==2.31.0
pyyaml==6.0.1
jsonschema==4.22.0
